feat(interface): use dynamic polymorphism to run different query implementations
