+++ src/core_server/internal/interface/backend.hpp
@@ -5,10 +5,10 @@
 #include <vector>
 
 #include "core_server/internal/coordination/catalog.hpp"
-#include "queries/simple_query.hpp"
 #include "core_server/internal/parsing/ceql_query/parser.hpp"
 #include "core_server/internal/stream/ring_tuple_queue/queue.hpp"
 #include "core_server/library/components/result_handler/result_handler.hpp"
+#include "queries/simple_query.hpp"
 #include "shared/datatypes/catalog/attribute_info.hpp"
 #include "shared/datatypes/catalog/event_info.hpp"
 #include "shared/datatypes/catalog/stream_info.hpp"
@@ -36,7 +36,7 @@ class Backend {
   uint64_t maximum_historic_time_between_events = 0;
   std::optional<uint64_t> previous_event_sent;
 
-  std::vector<std::unique_ptr<SimpleQuery<ResultHandlerT>>> queries;
+  std::vector<std::unique_ptr<GenericQuery<ResultHandlerT>>> queries;
   std::vector<Internal::ZMQMessageSender> inner_thread_event_senders = {};
 
  public:
@@ -100,7 +100,9 @@ class Backend {
     std::string inproc_receiver_address = "inproc://"
                                           + std::to_string(next_available_inproc_port++);
     queries.emplace_back(std::make_unique<SimpleQuery<ResultHandlerT>>(
-      std::move(parsed_query), catalog, queue, inproc_receiver_address, result_handler));
+      catalog, queue, inproc_receiver_address, result_handler));
+
+    queries.back()->init(std::move(parsed_query));
 
     query_events_time_window_mode.push_back(queries.back()->time_window.mode);
     query_events_expiration_time.emplace_back(queries.back()->time_of_expiration);
